fix: rename createCommand to createCliCommand.
Although this is breaking, it should be part of 1.0 release
