+++ src/test-util/index.ts
@@ -20,4 +20,3 @@ export * from './stringOptionCommand'
 export * from './setup'
 export * from './spyDisplay'
 export * from './verboseCommand'
-export { createCommand } from '../util'
