+++ src/test-util/setup.ts
@@ -4,7 +4,7 @@ import {
   // @ts-ignore
   CliCommandInstance
 } from '../CliCommand';
-import { createCommand } from '../util'
+import { createCliCommand } from '../createCliCommand'
 
 import { InMemoryPresenter } from './InMemoryPresenter'
 import { InMemoryPresenterFactory } from './InMemoryPresenterFactory';
@@ -24,7 +24,7 @@ export function createInMemoryCli(name: string, ...commands): Cli {
 export function setupCliCommandTest<Config, Context = {}>(command: CliCommand<Config, Context>, argv: string[], context: { config?: Config } & Context = {} as any) {
   const presenterFactory = new InMemoryPresenterFactory()
   const args = createCommandArgs(command, argv)
-  const cmd = createCommand(command, presenterFactory, context)
+  const cmd = createCliCommand(command, presenterFactory, context)
 
   return { cmd, args, argv, ui: cmd.ui as InMemoryPresenter }
 }
